update uint test

+++ contracts/library/HelperLib.sol
@@ -75,39 +75,39 @@ library HelperLib {
         return true;
     }
 
-    function isRight(bytes32 _hash, uint8 height) internal pure returns (bool) {
+    function isRight(bytes32 _hash, uint height) internal pure returns (bool) {
         return getBit(uint256(_hash), height);
     }
 
-    function setBit(bytes32 bitmap, uint8 index) internal pure returns (bytes32) {
+    function setBit(bytes32 bitmap, uint index) internal pure returns (bytes32) {
         return bytes32(uint256(bitmap) | (1 << (index & 0xff)));
     }
 
-    function setBit(uint256 bitmap, uint8 index) internal pure returns (uint256) {
+    function setBit(uint256 bitmap, uint index) internal pure returns (uint256) {
         return (bitmap | (1 << (index & 0xff)));
     }
 
-    function getBit(uint256 bitmap, uint8 index) internal pure returns (bool) {
+    function getBit(uint256 bitmap, uint index) internal pure returns (bool) {
         return ((bitmap & (1 << index)) > 0) ? true : false;
     }
 
-    function getBit(bytes32 bitmap, uint8 index) internal pure returns (bool) {
+    function getBit(bytes32 bitmap, uint index) internal pure returns (bool) {
         return ((uint256(bitmap) & (1 << index)) > 0) ? true : false;
     }
 
-    function clearBit(uint256 bitmap, uint8 index) internal pure returns (uint256) {
+    function clearBit(uint256 bitmap, uint index) internal pure returns (uint256) {
         return (bitmap & (~(1 << index)));
     }
 
-    function clearBit(bytes32 bitmap, uint8 index) internal pure returns (bytes32) {
+    function clearBit(bytes32 bitmap, uint index) internal pure returns (bytes32) {
         return bytes32(uint256(bitmap) & (~(1 << index)));
     }
 
-    function copyBits(bytes32 bitmap, uint8 index) internal pure returns (bytes32) {
+    function copyBits(bytes32 bitmap, uint index) internal pure returns (bytes32) {
         return bytes32((uint256(bitmap) >> index) << index);
     }
 
-    function parentPath(bytes32 path, uint8 height) internal pure returns (bytes32) {
+    function parentPath(bytes32 path, uint height) internal pure returns (bytes32) {
         if (height == 255) {
             return bytes32(0);
         }
